Refactoring + fixes on utilities and DetailedReservation class

+++ project/app/src/main/java/it/polito/mad/sportapp/Utilities.kt
@@ -464,14 +464,14 @@ fun dpToPx(context: Context, dp: Float): Int {
 
 /* time utilities */
 
-internal fun formatDuration(duration: Int): String {
+internal fun formatDuration(duration: Long): String {
 
     val hours = duration / 60
     val minutes = duration % 60
 
     return when {
-        duration < 60 -> "$duration" + "m"
-        minutes == 0 -> "$hours" + "h"
+        duration < 60L -> "$duration" + "m"
+        minutes == 0L -> "$hours" + "h"
         else -> "$hours" + "h" + " $minutes" + "m"
     }
 }
